Add an Errcheck plugin.
This is just a temporary hack.  The first problem is that TidyAll wants to
scan file names and errcheck wants to check packages.  So, the best
TidyAll can give you is a list of files in your repo.  You could argue
that it's possible to work your way back from these files and figure out
what the package is, but TidyAll opens the files into a temporary
directory, so you don't know much about the folder structure of the
place where the file actually lives.

I think TidyAll may need a validate_directory method. Or maybe it needs
the ability to set the command via argv and not require any matching
files to be found in order.  Or, maybe it needs the functionality to
pass args to the plugin, but not to the CLI itself.